[experimental][kleidi] move common functions to interface
